requirements.txt
streamlit~=1.36.0
requests~=2.32.3
python-dotenv~=1.0.1
PyPDF2~=3.0.1
langchain~=0.2.6
fastapi~=0.111.0